requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
huggingface_hub[hf_xet]

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# library to enable jyupter kernal
ipykernel

# Other libraries required for models execution
torch

sentence-transformers

# Library for UI
streamlit

# validation lib
pydantic
pydantic[email,timezone] 